Add files via upload
The program focuses on five main key topics from java:  if, else if, and else statements, loops, the PrintWriter class, the Scanner class and Array Lists,  the if statements  allow the user to decide if they want to create a list or not. Furthermore, the while and for loop allow the creation of the ShopNotes list content, if the while loop is true then, the items that are scanned by the Scanner class are added to the  ShopNotes' Array List content. Finally, thanks to the creation of a PrintWriter file, the ShopNotes ArrayList content is printed and stored as a text file.